# specs/001-epic-1-simple/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: FundFlow File Upload & Validation API
  description: API for uploading and validating PE fund distribution Excel files
  version: 1.2.0
  contact:
    name: FundFlow Development Team
servers:
  - url: http://localhost:8000/api
    description: Development server
paths:
  /upload:
    post:
      summary: Upload Excel file for processing
      description: Upload portfolio distribution Excel file, validate format and data, save to database
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Excel file (.xlsx or .xls) containing distribution data
              required:
                - file
      responses:
        '200':
          description: File uploaded and processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Invalid file or validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '413':
          description: File too large (>10MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported file type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    UploadResponse:
      type: object
      required:
        - session_id
        - filename
        - status
        - message
        - created_at
      properties:
        session_id:
          type: string
          format: uuid
          description: Unique session identifier
        filename:
          type: string
          description: Original filename
        status:
          $ref: '#/components/schemas/UploadStatus'
        message:
          type: string
          description: Status message
        created_at:
          type: string
          format: date-time
        file_hash:
          type: string
          description: SHA256 hash for duplicate detection
...
    ValidationError:
      type: object
      required:
        - row_number
        - column_name
        - error_code
        - error_message
        - severity
      properties:
        row_number:
          type: integer
          minimum: 1
        column_name:
          type: string
        error_code:
          type: string
          enum:
            - MISSING_HEADER
            - INVALID_FILENAME
            - EMPTY_FIELD
            - INVALID_ENTITY_TYPE
            - INVALID_STATE_CODE
            - NEGATIVE_AMOUNT
            - ZERO_DISTRIBUTIONS
            - DUPLICATE_INVESTOR
            - INVALID_NUMBER_FORMAT
            - ROW_LIMIT_EXCEEDED
            - FILE_SIZE_EXCEEDED
        error_message:
          type: string
          maxLength: 500
        severity:
          $ref: '#/components/schemas/ErrorSeverity'
        field_value:
          type: string
          nullable: true
    ValidationErrorResponse:
      type: object
      required:
        - session_id
        - status
        - errors
      properties:
        session_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/UploadStatus'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        summary:
          type: object
          properties:
            total_errors:
              type: integer
            error_count:
              type: integer
            warning_count:
              type: integer
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
        error_code:
          type: string
          nullable: true
    UploadStatus:
      type: string
      enum:
        - queued
        - uploading
        - parsing
        - validating
        - saving
        - completed
        - failed_upload
        - failed_parsing
        - failed_validation
        - failed_saving
...
    ErrorSeverity:
      type: string
      enum:
        - ERROR
        - WARNING
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session-based authentication for prototype
security:
  - sessionAuth: []
...